pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = "Веб-приложение для охотников 🫎 и рыболовов 🐟 "
authors = ["pequod <hounasamin1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
folium = "^0.15.1"
geos = "^0.2.3"
proj = "^0.2.0"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"

[tool.poetry.dev-dependencies]
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.0.0"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
include_trailing_comma = true
use_parentheses = true
line_length = 79
multi_line_output = 3
force_grid_wrap = 0

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git
    | \**/migrations
    | \**/__pychace__
    | \.venv
    | \/static
  )/
  | manage.py
)
'''
